Deduplicate keypair material accessors
