+++ crates/uselesskey-core/src/srp/keypair_material.rs
@@ -108,6 +108,87 @@ impl Pkcs8SpkiKeyMaterial {
     }
 }
 
+/// Implement the common PKCS#8/SPKI accessor and negative-fixture methods for a
+/// keypair fixture whose cached inner value stores a `material` field.
+///
+/// The caller must provide a type with an `inner.material` field containing a
+/// [`Pkcs8SpkiKeyMaterial`] and a `load_variant(&self, &str)` method whose
+/// returned inner value has the same `material` field.
+#[macro_export]
+macro_rules! impl_pkcs8_spki_fixture_accessors {
+    () => {
+        /// PKCS#8 DER-encoded private key bytes.
+        pub fn private_key_pkcs8_der(&self) -> &[u8] {
+            self.inner.material.private_key_pkcs8_der()
+        }
+
+        /// PKCS#8 PEM-encoded private key.
+        pub fn private_key_pkcs8_pem(&self) -> &str {
+            self.inner.material.private_key_pkcs8_pem()
+        }
+
+        /// SPKI DER-encoded public key bytes.
+        pub fn public_key_spki_der(&self) -> &[u8] {
+            self.inner.material.public_key_spki_der()
+        }
+
+        /// SPKI PEM-encoded public key.
+        pub fn public_key_spki_pem(&self) -> &str {
+            self.inner.material.public_key_spki_pem()
+        }
+
+        /// Write the PKCS#8 PEM private key to a tempfile and return the handle.
+        pub fn write_private_key_pkcs8_pem(
+            &self,
+        ) -> Result<$crate::sink::TempArtifact, $crate::Error> {
+            self.inner.material.write_private_key_pkcs8_pem()
+        }
+
+        /// Write the SPKI PEM public key to a tempfile and return the handle.
+        pub fn write_public_key_spki_pem(
+            &self,
+        ) -> Result<$crate::sink::TempArtifact, $crate::Error> {
+            self.inner.material.write_public_key_spki_pem()
+        }
+
+        /// Produce a corrupted variant of the PKCS#8 PEM.
+        pub fn private_key_pkcs8_pem_corrupt(&self, how: $crate::negative::CorruptPem) -> String {
+            self.inner.material.private_key_pkcs8_pem_corrupt(how)
+        }
+
+        /// Produce a deterministic corrupted PKCS#8 PEM using a variant string.
+        pub fn private_key_pkcs8_pem_corrupt_deterministic(&self, variant: &str) -> String {
+            self.inner
+                .material
+                .private_key_pkcs8_pem_corrupt_deterministic(variant)
+        }
+
+        /// Produce a truncated variant of the PKCS#8 DER.
+        pub fn private_key_pkcs8_der_truncated(&self, len: usize) -> Vec<u8> {
+            self.inner.material.private_key_pkcs8_der_truncated(len)
+        }
+
+        /// Produce a deterministic corrupted PKCS#8 DER using a variant string.
+        pub fn private_key_pkcs8_der_corrupt_deterministic(&self, variant: &str) -> Vec<u8> {
+            self.inner
+                .material
+                .private_key_pkcs8_der_corrupt_deterministic(variant)
+        }
+
+        /// Return a valid (parseable) public key that does *not* match this private key.
+        pub fn mismatched_public_key_spki_der(&self) -> Vec<u8> {
+            let other = self.load_variant("mismatch");
+            other.material.public_key_spki_der().to_vec()
+        }
+
+        /// A stable key identifier derived from the public key (base64url blake3 hash prefix).
+        #[cfg(feature = "jwk")]
+        pub fn kid(&self) -> String {
+            self.inner.material.kid()
+        }
+    };
+}
+
 const DEFAULT_KID_PREFIX_BYTES: usize = 12;
 
 fn kid_from_bytes(bytes: &[u8]) -> String {
